feature / change scss dir #1
