+++ convert_to_scss.js
@@ -1,0 +1,29 @@
+const fs = require('fs');
+const ejs = require('ejs');
+
+//テンプレ作成
+let template = "// Automatically generated by ./convert_to_scss.js\n\n$skin-name: \"<%- skin_name -%>\";\n$skin-colors: (<%- editor_colors -%>);";
+
+// ./src/skins/以下にdir名があるスキンの設定をSCSS化する.
+const filenames = fs.readdirSync('./src/skins/');
+filenames.forEach( (skin_name) => {
+    convert_to_scss(skin_name);
+})
+
+// 指定のスキン名のスキンの設定をSCSS化する.
+function convert_to_scss(skin_name) {
+    //skin.jsonを読み込み
+    const skin_json = JSON.parse(fs.readFileSync('./src/skins/' + skin_name + '/skin.json', 'utf8'));
+    const skin_obj = Object.keys(skin_json);//objectに変換
+    let editor_colors_txt = "";
+    const skin_name_txt = skin_json['name'];
+    const colors_obj = skin_json['settings']['color']['palette'];
+    colors_obj.forEach((color) => {
+        const color_txt = '"' + color['slug'] + '"' + ' : ' + color['color'] + ',\n';
+        console.log(color_txt);
+        editor_colors_txt += color_txt;
+    });
+    let scss = ejs.render(template, { skin_name: skin_name_txt, editor_colors: editor_colors_txt });
+    fs.writeFileSync('./src/skins/' + skin_name_txt + '/scss/_skin_config.scss', scss);
+    console.log(colors_obj.length + ' color has been set.');
+}
