+++ src/common/_variable.scss
@@ -1,5 +1,5 @@
 // --- skin prefix
-$prefix-skin--common: theidbrewing;
+$prefix-skin--common: tt1skin;
 
 // --- Colors
 // primary
