+++ src/common/components/_navigation.scss
@@ -33,7 +33,7 @@
                 display: flex;
                 width: var(--icon-size, 44px);
                 height: var(--icon-size, 44px);
-                background-image: svg-load("./build/images/icons/hamburger-menu.svg", fill=#{v.$color-accent-01-500});
+                background-image: svg-load("../../../../build/images/icons/hamburger-menu.svg", fill=#{v.$color-accent-01-500});
             }
         }
     }
