--- src/scss/style.scss
@@ -1,4 +1,0 @@
-@forward './config';
-@forward './components/logo';
-@forward './components/navigation';
-@forward './global';
