+++ src/skins/theidbrewing/scss/_skin_config.scss
@@ -1,0 +1,10 @@
+// Automatically generated by ./skin-config.js
+
+$skin-name: "theidbrewing";
+$skin-colors: ("primary" : #540CED,
+"body" : #1C0120,
+"base" : #FBFBFB,
+"accent-01" : #B900D8,
+"accent-02" : #ED2455,
+"accent-03" : #EDE12F,
+);
