+++ src/skins/theidbrewing/skin.json
@@ -1,0 +1,40 @@
+{
+    "name": "theidbrewing",
+    "settings": {
+        "google_fonts_url": "https://fonts.googleapis.com/css2?family=Noto+Sans+JP:wght@300;400;700&family=Open+Sans:wght@600&family=Space+Mono:wght@400;700&display=swap",
+        "color": {
+            "palette": [
+                {
+                    "name": "Primary",
+                    "slug": "primary",
+                    "color": "#540CED"
+                },
+                {
+                    "name": "Body",
+                    "slug": "body",
+                    "color": "#1C0120"
+                },
+                {
+                    "name": "Base",
+                    "slug": "base",
+                    "color": "#FBFBFB"
+                },
+                {
+                    "name": "Accent 01",
+                    "slug": "accent-01",
+                    "color": "#B900D8"
+                },
+                {
+                    "name": "Accent 02",
+                    "slug": "accent-02",
+                    "color": "#ED2455"
+                },
+                {
+                    "name": "Accent 03",
+                    "slug": "accent-03",
+                    "color": "#EDE12F"
+                }
+            ]
+        }
+    }
+}
